# diffeqmapreduce for faster anyeltypedual in some situations