--- a/src/forwarddiff.jl
+++ b/src/forwarddiff.jl
@@ -36,8 +36,21 @@ function promote_dual(::Type{T},
     ForwardDiff.Dual{T3, promote_dual(V, V2), N}
 end
 
-# Untyped dispatch: catch composite types, check all of their fields
+# `reduce` and `map` are specialized on tuples to be unrolled (via recursion)
+# Therefore, they can be type stable even with heterogenous input types.
+# We also don't care about allocating any temporaries with them, as it should
+# all be unrolled and optimized away.
+# Being unrolled also means const prop can work for things like
+# `mapreduce(f, op, propertynames(x))`
+# where `f` may call `getproperty` and thus have return type dependent
+# on the particular symbol.
+# `mapreduce` hasn't received any such specialization.
+@inline diffeqmapreduce(f::F, op::OP, x::Tuple) where{F,OP} = reduce(op, map(f, x))
+@inline diffeqmapreduce(f::F, op::OP, x::NamedTuple) where{F,OP} = reduce(op, map(f, x))
+# For other container types, we probably just want to call `mapreduce`
+@inline diffeqmapreduce(f::F, op::OP, x) where{F,OP} = mapreduce(f, op, x)
 
+# Untyped dispatch: catch composite types, check all of their fields
 """
     anyeltypedual(x)
 
@@ -64,8 +77,7 @@ function anyeltypedual(x, counter = 0)
     if propertynames(x) === ()
         Any
     elseif counter < 100
-        counter += 1
-        mapreduce(y -> !isdefined(x, y) ? Any : anyeltypedual(getproperty(x, y), counter),
+        diffeqmapreduce(y -> !isdefined(x, y) ? Any : anyeltypedual(getproperty(x, y), counter+1),
                   promote_dual, propertynames(x))
     else
         Any
@@ -153,19 +165,19 @@ function anyeltypedual(x::Tuple, counter = 0)
     if x === ()
         Any
     else
-        mapreduce(anyeltypedual, promote_dual, x)
+        diffeqmapreduce(anyeltypedual, promote_dual, x)
     end
 end
 function anyeltypedual(x::Dict, counter = 0)
     isempty(x) ? eltype(values(x)) : mapreduce(anyeltypedual, promote_dual, values(x))
 end
 function anyeltypedual(x::NamedTuple, counter = 0)
-    isempty(x) ? Any : mapreduce(anyeltypedual, promote_dual, values(x))
+    isempty(x) ? Any : diffeqmapreduce(anyeltypedual, promote_dual, values(x))
 end
-
-function promote_u0(u0, p, t0)
+@inline function promote_u0(u0, p, t0)
     if !(eltype(u0) <: ForwardDiff.Dual)
-        T = anyeltypedual(p)
+      T = anyeltypedual(p)
+      T === Any && return u0
         if T <: ForwardDiff.Dual
             return T.(u0)
         end