@chriselrod 2022-08-17
There are inference test failures.
I'll use the better/more type stable `Static.reduce_tup`.