@chriselrod 2022-08-17
On Core, it is now the new tests I added that are failing. They passed for me on 1.7.
I'll try 1.8 (is "1" now 1.8?) and master...