@chriselrod 2022-08-18
The failures were because of code cov adding a bunch of `Expr(:code_coverage_effect)`: https://github.com/SciML/DiffEqBase.jl/runs/7889809012?check_suite_focus=true