@chriselrod 2022-08-18
Is there any reason to think the failures are related?
```
 ERROR: LoadError: Package StochasticDiffEq errored during testing (received signal: KILL)
Stacktrace:
 [1] pkgerror(msg::String)
```
The failures certainly look bad.