@ChrisRackauckas 2022-08-18
Sounds like something odd has happened with the v1.8 update that we now have to look into. Fantastic. The allocator looks different: the OrdinaryDiffEq.jl code makes it looks like `similar` or `zero` on some array type has changed its output type.